[gnunet] Reduce the size of gnunet-service-ats.js
It was getting hard to debug such a huge file. The plugin size was
reduced by not linking in libraries and using symbols from the
main module which are now properly exported via -s MAIN_MODULE=1.
The size of the main module was reduced by preloading the plugin
instead of embedding it.